+++ js/gui/KeyboardShortcuts.js
@@ -41,8 +41,8 @@
 
     var nextFeed = function (navigationArea) {
         var current = navigationArea.find('.active');
-        var elements = navigationArea.find('.subscriptions:visible,' +
-                                           '.starred:visible,' +
+        var elements = navigationArea.find('.subscriptions-feed:visible,' +
+                                           '.starred-feed:visible,' +
                                            '.feed:visible');
 
         if (current.hasClass('folder')) {
@@ -74,8 +74,8 @@
 
     var previousFeed = function (navigationArea) {
         var current = navigationArea.find('.active');
-        var elements = navigationArea.find('.subscriptions:visible,' +
-                                           '.starred:visible,' +
+        var elements = navigationArea.find('.subscriptions-feed:visible,' +
+                                           '.starred-feed:visible,' +
                                            '.feed:visible');
 
         if (current.hasClass('folder')) {
@@ -90,7 +90,7 @@
             }
 
             // no subfeed found
-            var starred = $('.starred:visible');
+            var starred = $('.starred-feed:visible');
             if (starred.length > 0) {
                 starred.children('a:visible').trigger('click');
             }