+++ templates/part.navigation.unreadfeed.php
@@ -2,7 +2,7 @@
 		active: Navigation.isSubscriptionsActive(),
 		unread: Navigation.getUnreadCount() > 0
 	}"
-    class="subscriptions has-counter has-menu">
+    class="subscriptions-feed with-counter with-menu">
 
 	<a class="icon-rss" ng-href="#/items/" ng-if="!Navigation.isShowAll()">
 	   <?php p($l->t('Unread articles'))?>